fix: option TS parameters
